refactor: move private files from classes into internal
