+++ packages/classes/src/internal/custom-error.js
