+++ packages/classes/src/internal/file-mapping.js
