+++ packages/classes/src/internal/invalid-config-error.js
