+++ packages/classes/src/internal/local-assets.js
