+++ packages/classes/src/internal/local-file-location.js
