+++ packages/classes/src/internal/missing-config-error.js
