+++ packages/classes/src/internal/multiple-config-sources-error.js
