+++ packages/classes/src/internal/resolve-files.js
