+++ packages/classes/src/internal/resolved-files.js
