+++ packages/classes/src/internal/single-dest-multiple-source-error.js
