+++ packages/classes/test/resolved-files/iterator.test.js
@@ -1,5 +1,5 @@
 const tap = require('tap');
-const ResolvedFiles = require('../../src/resolved-files.js');
+const ResolvedFiles = require('../../src/internal/resolved-files.js');
 
 tap.test('ResolvedFiles: iterator for ./my/file.js', (t) => {
     const subject = new ResolvedFiles(['./my/file.js', './my/file.css'], {
